rename: 로그인 페이지 폴더 이름 변경
[#804]
